// cortex-js/src/infrastructure/dtos/threads/create-message.dto.ts
import { ChatCompletionRole } from '@/domain/models/message.interface';
import { ApiProperty } from '@nestjs/swagger';
import { IsEnum } from 'class-validator';

export class CreateMessageDto {
  @ApiProperty({
    description: `The role of the entity that is creating the message. Allowed values include:
      - user: Indicates the message is sent by an actual user and should be used in most cases to represent user-generated messages.
      - assistant: Indicates the message is generated by the assistant. Use this value to insert messages from the assistant into the conversation.`,
  })
  @IsEnum(ChatCompletionRole)
  role: ChatCompletionRole;

  @ApiProperty({
    description: 'The text contents of the message.',
  })
  content: string;
}